refactor(ui): split SpecifyAmountView out of its controller
The view was a `private struct` in the controller's tail, so it could not be
previewed and the file held both the SwiftUI screen and the UIKit host.

Two fixes came with the move. `EnterAmountView` was given `frame(minHeight: 90)`
— a floor, not a ceiling — and it ends in `frame(maxHeight: .infinity)`, so it
swallowed the `Spacer` below it and left the amount centred in the gap between
the title and the keypad. It gets the fixed 110 its own preview uses. The
controller's background was `SecondaryBackground` while the view had moved to
`primaryBackground`, which showed as a white strip in the safe areas.

Four previews: empty, an amount entered, dark, and accessibility type.
